Plot UserControl is seperated from WpfPlot User Control. Now WpfPlot is Just a User control that uses Plot() to display the grpah.
