+++ EasyPlot/WpfPlot.xaml
@@ -6,31 +6,7 @@
              xmlns:local="clr-namespace:EasyPlot"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-    <Grid Margin="10">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="Auto"/>
-            <ColumnDefinition Width="*"/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-        </Grid.RowDefinitions>
-        <TextBlock Margin="2" x:Name="tbTitle" Grid.Column="1" Grid.Row="0" RenderTransformOrigin="0.5,0.5" FontSize="14" FontWeight="Bold" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" TextAlignment="Center" Text="Title"/>
-        <TextBlock Margin="2" x:Name="tbDate" Grid.Column="0" Grid.ColumnSpan="2" Grid.Row="1" RenderTransformOrigin="0.5,0.5" FontSize="10" Foreground="Blue" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" TextAlignment="Left" Text="Now"/>
-        <TextBlock Margin="2" x:Name="tbResults" Grid.Column="1" Grid.Row="1" RenderTransformOrigin="0.5,0.5" FontSize="10" Foreground="Blue" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" TextAlignment="Right" Text="Results"/>
-        <TextBlock Margin="2" x:Name="tbXLabel" Grid.Column="1" Grid.Row="3" RenderTransformOrigin="0.5,0.5" TextAlignment="Center" Text="X Axis"/>
-        <TextBlock Margin="2" Name="tbYLabel" Grid.Column="0" Grid.Row="2" RenderTransformOrigin="0.5,0.5" TextAlignment="Center"  Text="Y Axis">
-            <TextBlock.LayoutTransform>
-                <RotateTransform Angle="-90"/>
-            </TextBlock.LayoutTransform>
-        </TextBlock>
-        <Grid Margin="0,0,0,0" x:Name ="chartGrid" Grid.Column="1" Grid.Row="2" ClipToBounds="False" Background="Transparent" SizeChanged="chartGrid_SizeChanged" >
-
-        </Grid>
-        <Canvas Margin="2" Name="textCanvas" Grid.Column="1" Grid.Row="2" ClipToBounds="True">
-            <Canvas Name="chartCanvas" ClipToBounds="True" Background="Transparent" MouseLeftButtonDown="OnMouseLeftButtonDown" MouseMove="OnMouseMove"  MouseLeftButtonUp="OnMouseLeftButtonUp" MouseRightButtonDown="OnMouseRightButtonDown" MouseWheel="OnMouseWheel" MouseLeave="OnMouseLeave" />
-        </Canvas>
+    <Grid x:Name="MainWindow_Grid">
+            
     </Grid>
 </UserControl>
